--- a/packages/grafana-plugin-configs/package.json
+++ b/packages/grafana-plugin-configs/package.json
@@ -13,6 +13,7 @@
     "fork-ts-checker-webpack-plugin": "8.0.0",
     "glob": "10.3.3",
     "replace-in-file-webpack-plugin": "1.0.6",
+    "swc-loader": "0.2.3",
     "webpack": "5.89.0"
   },
   "packageManager": "yarn@3.6.0"