--- a/packages/grafana-plugin-configs/webpack.config.ts
+++ b/packages/grafana-plugin-configs/webpack.config.ts
@@ -3,7 +3,7 @@ import ESLintPlugin from 'eslint-webpack-plugin';
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 import path from 'path';
 import ReplaceInFileWebpackPlugin from 'replace-in-file-webpack-plugin';
-import { Configuration } from 'webpack';
+import { Configuration, DefinePlugin } from 'webpack';
 
 import { DIST_DIR } from './constants';
 import { getPackageJson, getPluginJson, getEntries, hasLicense } from './utils';
@@ -29,12 +29,6 @@ const config = async (env: Record<string, unknown>): Promise<Configuration> => {
         config: [__filename],
       },
       cacheDirectory: path.resolve(__dirname, '../../.yarn/.cache/webpack', path.basename(process.cwd())),
-      cacheLocation: path.resolve(
-        __dirname,
-        '../../.yarn/.cache/eslint-webpack-plugin',
-        path.basename(process.cwd()),
-        '.eslintcache'
-      ),
     },
 
     context: process.cwd(),
@@ -186,23 +180,35 @@ const config = async (env: Record<string, unknown>): Promise<Configuration> => {
           ],
         },
       ]),
-      new ForkTsCheckerWebpackPlugin({
-        async: Boolean(env.development),
-        issue: {
-          include: [{ file: '**/*.{ts,tsx}' }],
-        },
-        typescript: { configFile: path.join(process.cwd(), 'tsconfig.json') },
-      }),
-      new ESLintPlugin({
-        extensions: ['.ts', '.tsx'],
-        lintDirtyModulesOnly: Boolean(env.development), // don't lint on start, only lint changed files
-      }),
+      env.development
+        ? new ForkTsCheckerWebpackPlugin({
+            async: true,
+            issue: {
+              include: [{ file: '**/*.{ts,tsx}' }],
+            },
+            typescript: { configFile: path.join(process.cwd(), 'tsconfig.json') },
+          })
+        : new DefinePlugin({}),
+      env.development
+        ? new ESLintPlugin({
+            extensions: ['.ts', '.tsx'],
+            lintDirtyModulesOnly: true, // don't lint on start, only lint changed files
+            cacheLocation: path.resolve(
+              __dirname,
+              '../../.yarn/.cache/eslint-webpack-plugin',
+              path.basename(process.cwd()),
+              '.eslintcache'
+            ),
+          })
+        : new DefinePlugin({}),
     ],
 
     resolve: {
-      extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      extensions: ['.ts', '.tsx', '.js', '.jsx'],
       unsafeCache: true,
     },
+
+    stats: env.production ? 'errors-only' : 'normal',
 
     watchOptions: {
       ignored: ['**/node_modules', '**/dist', '**/.yarn'],