--- a/public/app/plugins/datasource/grafana-testdata-datasource/package.json
+++ b/public/app/plugins/datasource/grafana-testdata-datasource/package.json
@@ -26,7 +26,6 @@
     "@types/node": "18.18.5",
     "@types/react": "18.2.15",
     "@types/testing-library__jest-dom": "5.14.8",
-    "swc-loader": "0.2.3",
     "ts-node": "10.9.1",
     "webpack": "5.89.0"
   },